pyproject.toml
[project]
name = "oversight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.0",
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.33",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=0.6.7",
    "openreview-py>=1.52.2",
    "pgvector>=0.4.1",
    "psycopg>=3.2.10",
    "pymongo>=4.15.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "sickle>=0.7.0",
    "tqdm>=4.67.1",
    "wheel>=0.45.1",
    "xmltodict>=1.0.2",
    "yagmail>=0.15.293",
    "agentica",
    "beautifulsoup4>=4.14.3",
    "semanticscholar>=0.12.0",
]

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
superscrape = "superscraper.cli:entrypoint"

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "ty>=0.0.1a7",
]

[tool.ruff]
exclude = ["src/agentica-mini"]

[tool.ty.src]
exclude = ["src/agentica-mini"]

[tool.uv.sources]
agentica = { path = "./src/agentica-mini", editable = true }
